updated openai, assistants wip

+++ math_rag/application/models/inference/em_status_tracker.py
@@ -8,4 +8,4 @@ class EMStatusTracker(BaseModel):
     num_tasks_failed: int = 0
     num_rate_limit_errors: int = 0
     num_api_errors: int = 0
-    time_of_last_rate_limit_error: int = 0
+    time_of_last_rate_limit_error: float = 0.0
