+++ math_rag/application/models/inference/llm_status_tracker.py
@@ -9,4 +9,4 @@ class LLMStatusTracker(BaseModel):
     num_rate_limit_errors: int = 0
     num_api_errors: int = 0
     num_non_api_errors: int = 0
-    time_of_last_rate_limit_error: int = 0
+    time_of_last_rate_limit_error: float = 0.0
